- move clirunner utils under ./cli directory
