+++ lib/cli/clirunner.js
@@ -3,10 +3,20 @@
 const CodeRunner    = require('../index'),
       program       = require('commander'),
       logger        = require('../util/logger'),
-      terminal      = require('../util/terminal'),
+      utils         = require('./utils'),
+      async         = require('../util/async'),
       packageConfig = require(`${__dirname}/../../package.json`);
 
 const getRunOptions = require('./options');
+
+const modelWarningMsg = modelName => `IMPORTANT!
+The business logic code will be deployed to model "${ modelName }".
+Any business logic which is already deployed on the server in that model
+will be removed and replaced with the code from your current project.`;
+
+const modelConfirmMsg = `If this is an undesired behavior, stop now and set a different deployment model
+either by using the --model argument or changing the model name in coderunner.json.
+Would you like to continue? (Y/N)`;
 
 process.on('unhandledRejection', r => logger.error(r));
 
@@ -29,15 +39,15 @@ const setTerminationHook = runner => {
 
 const startRunner = ({ appRequired, repoPathRequired, createRunner, showVersion, showCopyright, showModelConfirm }) => {
 
-  showVersion && terminal.printVersion();
-  showCopyright && terminal.printCopyright();
+  showVersion && printVersion();
+  showCopyright && printCopyright();
 
   const confirmModelIfNeeded = opts => {
     if (!showModelConfirm) {
       return opts;
     }
 
-    return terminal.showModelAlert(opts.app.model).then(() => opts);
+    return showModelAlert(opts.app.model).then(() => opts);
   };
 
   return Promise.resolve()
@@ -126,6 +136,14 @@ if (!program.args.length) {
   debug();
 }
 
+function printVersion() {
+  logger.info(`CodeRunner(tm) Backendless JavaScript CodeRunner v${packageConfig.version}`);
+}
+
+function printCopyright() {
+  logger.info(`Copyright(C) ${new Date().getFullYear()} Backendless Corp. All rights reserved. `);
+}
+
 function exit() {
   process.exit();
 }
@@ -135,3 +153,21 @@ function exitWithError(err) {
 
   process.exit();
 }
+
+const showModelAlert = async(function* (modelName) {
+  logger.info(modelWarningMsg(modelName));
+
+  if (program.quiet) {
+    return;
+  }
+
+  const confirmed = (yield utils.confirmation(modelConfirmMsg));
+
+  if (confirmed === false) {
+    process.exit();
+  }
+
+  if (confirmed === undefined) {
+    return showModelAlert(modelName);
+  }
+});
