+++ lib/cli/program.js
@@ -3,7 +3,7 @@
 const program = require('commander');
 const logger = require('../util/logger');
 const async = require('../util/async');
-const terminal = require('../util/terminal');
+const utils = require('./utils');
 
 const UUID_PATTERN = '[A-F0-9\\-]{36}';
 const DEFAULT_REPO_PATH = '/var/lib/backendless/repo/';
@@ -43,8 +43,8 @@ function ensureApp(options) {
   app.model = program.model || app.model || DEFAULT_MODEL;
 
   const gatherAppOptions = async(function* () {
-    app.id = program.appId || app.id || (yield terminal.promptParam('Application ID'));
-    app.apiKey = program.appKey || app.apiKey || app.secretKey || (yield terminal.promptParam('API Key'));
+    app.id = program.appId || app.id || (yield promptParam('Application ID'));
+    app.apiKey = program.appKey || app.apiKey || app.secretKey || (yield promptParam('API Key'));
   });
 
   return gatherAppOptions()
@@ -60,4 +60,8 @@ function assertValueIsUUID(value, name) {
   if (!value || !value.match(UUID_PATTERN)) {
     throw new Error(`${name} is invalid`);
   }
+}
+
+function promptParam(name) {
+  return utils.prompt(`Please enter ${ name } and press [Enter]:`);
 }
